requirements.txt
python-dotenv>=1.0.1
openai>=1.40.0
tiktoken>=0.7.0
chromadb>=0.5.3
sentence-transformers>=3.0.0
pypdf>=4.2.0
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.8.4
scikit-learn>=1.5.1
streamlit>=1.37.0
pyyaml>=6.0.2
tqdm>=4.66.0
rich>=13.7.1
